Refactor userTest into separate test files
